Rename grav_d.py to gravd.py
